+++ backend/app/models/__init__.py
@@ -1,0 +1,4 @@
+from .item import *
+from .msg import *
+from .user import *
+from sqlmodel import SQLModel
