+++ backend/app/models/base.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+import uuid
+
+from pydantic import EmailStr
+from sqlmodel import Field, Relationship, SQLModel
+from sqlalchemy.orm import Mapped
